Move DdmlibToLog4jWrapper and AdbException to the appropriate package
Issue: #196
