+++ src/name/mlopatkin/andlogview/device/AdbException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 Mikhail Lopatkin
+ * Copyright 2022 the Andlogview authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package name.mlopatkin.andlogview.liblogcat.ddmlib;
+package name.mlopatkin.andlogview.device;
 
 public class AdbException extends Exception {
     public AdbException() {
