+++ src/name/mlopatkin/andlogview/device/DdmlibToLog4jWrapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 Mikhail Lopatkin
+ * Copyright 2022 the Andlogview authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package name.mlopatkin.andlogview.liblogcat.ddmlib;
+package name.mlopatkin.andlogview.device;
 
 import com.android.ddmlib.Log.ILogOutput;
 import com.android.ddmlib.Log.LogLevel;
 
 import org.apache.log4j.Logger;
 
-public class DdmlibToLog4jWrapper implements ILogOutput {
+class DdmlibToLog4jWrapper implements ILogOutput {
     private static final Logger logger = Logger.getLogger("DDMLIB");
 
     @Override
